requirements.txt
fastapi
uvicorn[standard]
httpx
pandas
aiohttp
mcp-proxy
fastapi-mcp
pillow
openpyxl
authlib
fastapi-cache2
aioredis
fastapi-limiter
redis
python-jose[cryptography]
passlib[bcrypt]